requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-secret-manager
psycopg2